Hotel-Booking-Agent/requirements.txt
streamlit
playwright
html2text
crewai
python-dotenv